+++ src/chains/ethereum/ethereum.ts
@@ -8,9 +8,15 @@ import { TokenValue } from '../../services/base';
 
 // MKR does not match the ERC20 perfectly so we need to use a separate ABI.
 const MKR_ADDRESS = '0x9f8f72aa9304c8b593d555f12ef6589cc3a579a2';
-const UNISWAP_ROUTER = '0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D';
-const UNISWAP_V3_ROUTER = '0xE592427A0AEce92De3Edee1F18E0157C05861564';
-const UNISWAP_V3_NFT_MANAGER = '0xC36442b4a4522E871399CD717aBDD847Ab11FE88';
+
+// Balancer addresses
+const BALANCER_MAINNET_ADDRESS = '0x3E66B66Fd1d0b02fDa6C811Da9E0547970DB2f21';
+const BALANCER_KOVAN_ADDRESS = '0x4e67bf5bD28Dd4b570FBAFe11D0633eCbA2754Ec';
+
+// Uniswap addresses
+const UNISWAP_ROUTER_ADDRESS = '0x7a250d5630B4cF539739dF2C5dAcb4c659F2488D';
+const UNISWAP_V3_ROUTER_ADDRESS = '0xE592427A0AEce92De3Edee1F18E0157C05861564';
+const UNISWAP_V3_NFT_MANAGER_ADDRESS = '0xC36442b4a4522E871399CD717aBDD847Ab11FE88';
 
 export class Ethereum extends EthereumBase {
   private ethGasStationUrl: string;
@@ -19,11 +25,13 @@ export class Ethereum extends EthereumBase {
   private _approvedSpenders: string[];
 
   constructor() {
-    let config;
+      let config, otherApprovedSpenders;
     if (ConfigManager.config.ETHEREUM_CHAIN === 'mainnet') {
-      config = EthereumConfig.config.mainnet;
+        config = EthereumConfig.config.mainnet;
+        otherApprovedSpenders = [BALANCER_MAINNET_ADDRESS];
     } else {
-      config = EthereumConfig.config.kovan;
+        config = EthereumConfig.config.kovan;
+        otherApprovedSpenders = [BALANCER_KOVAN_ADDRESS];
     }
 
     super(
@@ -41,10 +49,10 @@ export class Ethereum extends EthereumBase {
     this.gasPrice = ConfigManager.config.ETH_MANUAL_GAS_PRICE;
     this.gasPriceLastUpdated = null;
     this._approvedSpenders = [
-      UNISWAP_ROUTER,
-      UNISWAP_V3_ROUTER,
-      UNISWAP_V3_NFT_MANAGER,
-    ];
+      UNISWAP_ROUTER_ADDRESS,
+      UNISWAP_V3_ROUTER_ADDRESS,
+      UNISWAP_V3_NFT_MANAGER_ADDRESS,
+    ].concat( otherApprovedSpenders);
 
     this.updateGasPrice();
   }